Use one single data generation entrypoint.
