--- reference/latest/src/client/java/com/example/docs/datagen/FabricDocsReferenceAdvancementGenerator.java
@@ -1,14 +1,0 @@
-package com.example.docs.datagen;
-
-import net.fabricmc.fabric.api.datagen.v1.DataGeneratorEntrypoint;
-import net.fabricmc.fabric.api.datagen.v1.FabricDataGenerator;
-
-// :::datagen-advancements:generator
-public class FabricDocsReferenceAdvancementGenerator implements DataGeneratorEntrypoint {
-	@Override
-	public void onInitializeDataGenerator(FabricDataGenerator fabricDataGenerator) {
-		FabricDataGenerator.Pack pack = fabricDataGenerator.createPack();
-		pack.addProvider(FabricDocsReferenceAdvancementProvider::new);
-	}
-}
-// :::datagen-advancements:generator
