+++ reference/latest/src/client/java/com/example/docs/datagen/FabricDocsReferenceDataGenerator.java
@@ -1,9 +1,16 @@
 package com.example.docs.datagen;
 
-// :::datagen-setup:generator
+import static com.example.docs.datagen.FabricDocsReferenceDamageTypesProvider.TATER_DAMAGE_TYPE;
+
+import net.minecraft.registry.RegistryBuilder;
+import net.minecraft.registry.RegistryKeys;
+
 import net.fabricmc.fabric.api.datagen.v1.DataGeneratorEntrypoint;
 import net.fabricmc.fabric.api.datagen.v1.FabricDataGenerator;
 
+import com.example.docs.damage.FabricDocsReferenceDamageTypes;
+
+// :::datagen-setup:generator
 public class FabricDocsReferenceDataGenerator implements DataGeneratorEntrypoint {
 	@Override
 	public void onInitializeDataGenerator(FabricDataGenerator fabricDataGenerator) {
@@ -11,7 +18,34 @@ public class FabricDocsReferenceDataGenerator implements DataGeneratorEntrypoint
 		// :::datagen-setup:pack
 		FabricDataGenerator.Pack pack = fabricDataGenerator.createPack();
 		// :::datagen-setup:pack
+
+		pack.addProvider(EnchantmentGenerator::new);
+
+		pack.addProvider(FabricDocsReferenceAdvancementProvider::new);
+
+		pack.addProvider(FabricDocsReferenceEnglishLangProvider::new);
+
+		pack.addProvider(FabricDocsReferenceItemTagProvider::new);
+
+		pack.addProvider(FabricDocsReferenceRecipeProvider::new);
+
+		pack.addProvider(FabricDocsReferenceBlockLootTableProvider::new);
+		pack.addProvider(FabricDocsReferenceChestLootTableProvider::new);
+
+		pack.addProvider(FabricDocsReferenceDamageTypesProvider.TaterDamageTypesGenerator::new);
+		pack.addProvider(FabricDocsReferenceDamageTypesProvider.TaterDamageTypeTagGenerator::new);
+
 		// :::datagen-setup:generator
 	}
+
+	// :::datagen-setup:generator
+	@Override
+	public void buildRegistry(RegistryBuilder registryBuilder) {
+		registryBuilder.addRegistry(RegistryKeys.DAMAGE_TYPE, registerable -> {
+			registerable.register(FabricDocsReferenceDamageTypes.TATER_DAMAGE, TATER_DAMAGE_TYPE);
+		});
+	}
+
+	// :::datagen-setup:generator
 }
 // :::datagen-setup:generator
