--- reference/latest/src/client/java/com/example/docs/datagen/FabricDocsReferenceEnchantmentGenerator.java
@@ -1,13 +1,0 @@
-package com.example.docs.datagen;
-
-import net.fabricmc.fabric.api.datagen.v1.DataGeneratorEntrypoint;
-import net.fabricmc.fabric.api.datagen.v1.FabricDataGenerator;
-
-public class FabricDocsReferenceEnchantmentGenerator implements DataGeneratorEntrypoint {
-	//#initdatagen
-	@Override
-	public void onInitializeDataGenerator(FabricDataGenerator fabricDataGenerator) {
-		FabricDataGenerator.Pack pack = fabricDataGenerator.createPack();
-		pack.addProvider(EnchantmentGenerator::new);
-	}
-}
