--- reference/latest/src/client/java/com/example/docs/datagen/FabricDocsReferenceLootTableGenerator.java
@@ -1,18 +1,0 @@
-package com.example.docs.datagen;
-
-import net.fabricmc.fabric.api.datagen.v1.DataGeneratorEntrypoint;
-import net.fabricmc.fabric.api.datagen.v1.FabricDataGenerator;
-
-// :::datagen-loot-tables:generator
-public class FabricDocsReferenceLootTableGenerator implements DataGeneratorEntrypoint {
-	@Override
-	public void onInitializeDataGenerator(FabricDataGenerator fabricDataGenerator) {
-		FabricDataGenerator.Pack pack = fabricDataGenerator.createPack();
-		// Add providers here
-		// :::datagen-loot-tables:generator
-		pack.addProvider(FabricDocsReferenceBlockLootTableProvider::new);
-		pack.addProvider(FabricDocsReferenceChestLootTableProvider::new);
-		// :::datagen-loot-tables:generator
-	}
-}
-// :::datagen-loot-tables:generator
