--- reference/latest/src/client/java/com/example/docs/datagen/FabricDocsReferenceRecipeGenerator.java
@@ -1,14 +1,0 @@
-package com.example.docs.datagen;
-
-// :::datagen-recipes:generator
-import net.fabricmc.fabric.api.datagen.v1.DataGeneratorEntrypoint;
-import net.fabricmc.fabric.api.datagen.v1.FabricDataGenerator;
-
-public class FabricDocsReferenceRecipeGenerator implements DataGeneratorEntrypoint {
-	@Override
-	public void onInitializeDataGenerator(FabricDataGenerator fabricDataGenerator) {
-		FabricDataGenerator.Pack pack = fabricDataGenerator.createPack();
-		pack.addProvider(FabricDocsReferenceRecipeProvider::new);
-	}
-}
-// :::datagen-recipes:generator
